Add tests to mp

+++ packages/mp/tests/test_mp/test_build_project/test_marketplace/test_deconstruct_integration.py
@@ -61,6 +61,18 @@ def test_deconstruct_built_integration(
         assert_deconstruct_integration(built_integration)
 
 
+def test_non_existing_integration_raises_file_not_found_error(
+    tmp_path: pathlib.Path,
+    mock_get_marketplace_path: str,
+    assert_deconstruct_integration: Callable[[pathlib.Path], None],
+) -> None:
+    with (
+        unittest.mock.patch(mock_get_marketplace_path, return_value=tmp_path),
+        pytest.raises(FileNotFoundError, match="Invalid integration .*"),
+    ):
+        assert_deconstruct_integration(tmp_path / "fake_integration")
+
+
 @pytest.fixture
 def assert_deconstruct_integration(
     tmp_path: pathlib.Path,
@@ -70,8 +82,11 @@ def assert_deconstruct_integration(
         commercial: pathlib.Path = tmp_path / mp.core.constants.COMMERCIAL_DIR_NAME
         shutil.copytree(integration_path.parent, commercial)
         integration: pathlib.Path = commercial / integration_path.name
-        requirements: pathlib.Path = integration / mp.core.constants.REQUIREMENTS_FILE
-        requirements.write_text("requests==2.32.3\n", encoding="utf-8")
+        if integration.exists():
+            requirements: pathlib.Path = (
+                integration / mp.core.constants.REQUIREMENTS_FILE
+            )
+            requirements.write_text("requests==2.32.3\n", encoding="utf-8")
 
         marketplace: Marketplace = mp.build_project.marketplace.Marketplace(commercial)
         marketplace.deconstruct_integration(integration)
